move the test to bountycastle
